// disney-characters/src/app/shared/interfaces/character.interface.ts
export interface CharacterInterface {
    _id: number;
    films: Array<any>;
    shortFilms: Array<any>;
    tvShows: Array<any>;
    videoGames: Array<any>;
    parkAttractions: Array<any>;
    allies: Array<any>;
    enemies: Array<any>;
    sourceUrl: string;
    name: string;
    imageUrl: string;
    createdAt: string;
    updatedAt: string;
    url: string;
    __v: number;
  }